+++ .env.example
@@ -14,6 +14,10 @@ DB_LOCAL_URI=mongodb://user-service-db:27017/peerprepUserServiceDB
 PORT=8003
 ENV=DEV # or PROD to use cloud URI
 
+# History Service
+HISTORY_SERVICE_PORT=8090
+HISTORY_SERVICE_MONGODB_URI=
+
 # Kafka Cluster
 KAFKA_BROKER_URI=kafka:8005
